Moved the console logic into main
These two functions had overlapping behaviour so it made sense to merge them.
Streamlined the logging setup
